Big directory restructure; need to fix tests
