+++ __tests__/components/AbacusHeader.test.tsx
@@ -1,11 +1,11 @@
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import { mantineRecoilWrap } from '../helpers/testHelpers';
-import AbacusHeader from '../../components/AbacusHeader';
+import AppHeader from '../../components/utils/AppHeader';
 
 describe('AbacusHeader', () => {
   it('renders a heading with title and theme switcher', () => {
-    render(mantineRecoilWrap(<AbacusHeader />));
+    render(mantineRecoilWrap(<AppHeader />));
 
     const heading = screen.getByRole('heading', {
       name: /FQM Testify: an ECQM Analysis Tool/i
