+++ __tests__/components/MeasureUpload.test.tsx
@@ -3,7 +3,7 @@ import '@testing-library/jest-dom';
 import { Calculator } from 'fqm-execution';
 import { mantineRecoilWrap, getMockRecoilState } from '../helpers/testHelpers';
 import { measureBundleState } from '../../state/atoms/measureBundle';
-import MeasureUpload from '../../components/MeasureUpload';
+import MeasureUpload from '../../components/measure-upload/MeasureUpload';
 
 const MOCK_BUNDLE: fhir4.Bundle = {
   resourceType: 'Bundle',
